Merged in fix/UD-161-security---user-is-able-to-access (pull request #331)
UD-161/UD-525/UD-577 - Protect admin routes with relevant permissions

* add ProtectedRoute & refactor usePermission

* feat: rename and move Permission functions

* feat: update usePermission to accept PermissionCondition

Enables OR conditions
Could also be expanded in future to support nesting (i.e.
`PermissionCondition.permissions` is an array of `PermissionId`s and/or
other `PermissionCondition`s)

* fix: UD-577 - no redirect on no view_school perm

Instead of a full redirect, we can display 0 rows in the SchoolTable,
when the User doesn't have the view_school perm

* feat: nested PermissionOption in usePermission

For the <ProtectedRoute/> component, we need complex AND/OR permission
combinations, which need an object to represent that can be recursively
evaluated

* feat: apply ProtectedRoute to all Admin pages

* feat: hide /admin/schools if missing define_ perm

* test: ProtectedRoute

* feat: omit `render` prop from ProtectedRoute

Custom render method, would override passed prop.
No current need for render, but would need to decide whether it took
priority over children

* refactor: PermissionCondition to use flat `AND/OR`


Approved-by: Henrik Wassdahl